Fix tracker data bug
Hooks only causes a re-render when the instance changes. Because I was modifying the data but using the same instance, the data would be updated but the app wouldn't be re-rendered. One way to fix this was to change the paradigm of how hooks is being used for tracker data. Now instead of updating tracker data, we just say "tracker data has been updated. please re-render" and pass it in an empty object (which is a new instance and forces a re-render).

I'm not sure if I am a fan of this paradigm but I think it's worth a try for this project so I can learn the pros and cons.